requirements.txt
# Core dependencies
snakemake==8.29
polars==0.20.6
requests==2.31.0
pyyaml==6.0.1
pyarrow==14.0.1
duckdb-engine==0.9.2
ibis-framework[duckdb]==7.1.0
graphviz==0.20.1

# Dev dependencies
ruff>=0.2.0
black>=24.1.0
mypy>=1.8.0